refactor: split config reply dispatch
